pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.7",
    "imap-tools>=1.10.0",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
